Implemented clean formatting and idempotency for tasks...
The clean formatting I believe makes this playbook easier to
read and cleaner. Also implemented stat checks in order to
ensure make tasks do not execute everytime the playbook is
executed to ensure idempotency. It may be desirable to implement
re-run make tasks by leveraging var(s) in order to re-make and
compile code and etc.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.coml>