Addressed #23 requests of changes
Changed check file var locations from /var/log due to permission issues.
Created gourl var to use for get_url…
The above address #23

Also in addition to the above added the below.
Sorted vars alphabetically to cleanup.
Created var for go tarball to use as a var for each tasks. In case this
ever needs to change and to follow consistently.