Total Revamp and Improvement to Documentation (#94)
This PR introduces a slew of badly needed improvements to the documentation. It not only adds more information that was previously missing, but it also restructures the documentation to be more readable, and friendly to the way a new user might approach the docs.

This PR does NOT complete the documentation. There are still a number of areas that need more information (indicated by missing sections and a "TBD" message). However, it does introduce a LOT of work towards this end, and more importantly improves the overall structure as mentioned before. With this new structure, it's more obvious what's missing, and where to find information that has been documented.